% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bridge}
\alias{bridge}
\title{Make a REST call to Bridge}
\usage{
bridge(endpoint, method, debug = FALSE, ...)
}
\arguments{
\item{endpoint}{The API endpoint.}

\item{method}{An HTTP verb function.}

\item{debug}{Whether to print the request string and body. FALSE by default.}

\item{...}{named arguments to pass to \code{httr::modify_url}.}
}
\description{
Make a REST call to Bridge
}
