% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bridgeDELETE}
\alias{bridgeDELETE}
\title{Make a DELETE call to Bridge}
\usage{
bridgeDELETE(endpoint, debug = FALSE, ...)
}
\arguments{
\item{endpoint}{Server endpoint.}

\item{debug}{Whether to print the request string and body. FALSE by default.}

\item{...}{Either the body of the request or named arguments to
pass to \code{httr::modify_url}.}
}
\description{
Make a DELETE call to Bridge
}
