% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bridgePOST}
\alias{bridgePOST}
\title{Make a POST call to Bridge}
\usage{
bridgePOST(endpoint, ...)
}
\arguments{
\item{endpoint}{Server endpoint.}

\item{...}{named arguments to pass to \code{httr::modify_url}}
}
\description{
Make a POST call to Bridge
}
