% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participants.R
\name{create_participant}
\alias{create_participant}
\title{Create a Bridge account}
\usage{
create_participant(
  phone_number = NULL,
  email = NULL,
  study = NULL,
  external_id = NULL,
  data_groups = NULL,
  phone_region_code = "US",
  sharing_scope = "all_qualified_researchers"
)
}
\arguments{
\item{phone_number}{The phone number associated with the account.}

\item{email}{The email address associated with the account.}

\item{study}{The study identifier.}

\item{external_id}{The participant identifier.}

\item{data_groups}{One or more data groups to assign this user to.}

\item{phone_region_code}{The CLDR two-letter region code describing the
region in which the phone number was issued. By default: "US".}

\item{sharing_scope}{One of "no_sharing", "sponsors_and_partners",
or "all_qualified_researchers". By default: "all_qualified_researchers".}
}
\description{
Creates a participant account tied to a phone number.
}
