% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participants.R
\name{delete_participant}
\alias{delete_participant}
\title{Delete a Bridge account}
\usage{
delete_participant(health_code = NULL, user_id = NULL)
}
\arguments{
\item{health_code}{The healthcode of the account to delete.}

\item{user_id}{The user ID of the account to delete.}
}
\description{
Only users with a \code{test_user} data group may be deleted.
}
