% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_events.R
\name{get_activity_events}
\alias{get_activity_events}
\title{Get activity events of a study participant (user) using their ID}
\usage{
get_activity_events(user_id)
}
\arguments{
\item{user_id}{The user ID of the participant}
}
\description{
Get activity events of a study participant (user) using their ID
}
