% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adherence.R
\name{get_adherence}
\alias{get_adherence}
\title{Endpoint to gather user adherence from Bridge}
\usage{
get_adherence(study_id, user_id)
}
\arguments{
\item{study_id}{the Bridge Study ID}

\item{user_id}{the Bridge User ID}
}
\value{
key-value-pair of adherence
}
\description{
Endpoint to gather user adherence from Bridge
}
