% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollments.R
\name{get_all_enrollments}
\alias{get_all_enrollments}
\title{Get enrollments for this study}
\usage{
get_all_enrollments(study_id, filter = "all", include_testers = FALSE)
}
\arguments{
\item{study_id}{The study ID.}

\item{filter}{One of "enrolled", "withdrawn", or "all". Defaults to "all".}

\item{include_testers}{Whether to include accounts with the
"test_user" data group.}
}
\value{
A list of accounts that are enrolled in that study.
}
\description{
Get enrollments for this study
}
