% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participants.R
\name{get_all_participants}
\alias{get_all_participants}
\title{Get all participants in the study}
\usage{
get_all_participants()
}
\value{
A list of participant accounts
}
\description{
This users the deprecated GET /v3/participants API and may stop working
at any time.
}
