% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies.R
\name{get_studies}
\alias{get_studies}
\title{Function to get studies of Bridge App}
\usage{
get_studies()
}
\value{
response from Bridge /studies endpoint
}
\description{
Function to get studies of Bridge App
}
