% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participants.R
\name{search_participants}
\alias{search_participants}
\title{Search for a study participant}
\usage{
search_participants(email = NULL, phone = NULL, study_id = NULL)
}
\arguments{
\item{email}{The email address to filter upon.}

\item{phone}{The phone number to filter upon.}

\item{study_id}{The study ID. Will use the v5 APIs (rather than v3).}
}
\value{
The first 100 search results.
}
\description{
Returns the first 100 results.
}
