% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bridge}
\alias{bridge}
\title{Make a REST call to Bridge}
\usage{
bridge(endpoint, method, body = NULL, ...)
}
\arguments{
\item{endpoint}{Server endpoint.}

\item{method}{An HTTP verb function}

\item{body}{The body of the request}

\item{...}{named arguments to pass to \code{httr::modify_url}}
}
\description{
Make a REST call to Bridge
}
