% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bridge_login}
\alias{bridge_login}
\title{Log in to Bridge}
\usage{
bridge_login(study, email = NULL, password = NULL)
}
\arguments{
\item{email}{Your email}

\item{password}{Your Bridge password}
}
\description{
Log in by inputting your username and password or cache your credentials in
a YAML file with the name \code{.bridgeCredentials} in your home (~) directory.
This file should have keys \code{email} and \code{password}.
}
