% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bridgeGET}
\alias{bridgeGET}
\title{Make a GET call to Bridge}
\usage{
bridgeGET(endpoint, ...)
}
\arguments{
\item{endpoint}{Server endpoint.}

\item{...}{named arguments to pass to \code{httr::modify_url}}
}
\description{
Make a GET call to Bridge
}
