% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participants.R
\name{get_participant}
\alias{get_participant}
\title{Get a study participant (user) record using the external ID of the account}
\usage{
get_participant(external_id)
}
\arguments{
\item{external_id}{The external ID of the participant}
}
\description{
Get a study participant (user) record using the external ID of the account
}
